refactor: Migrate to es6; Add webpack
Rewrite code to es6; change project structure

#19
