+++ src/customCommands.js
@@ -1,0 +1,3 @@
+const myCustomCommands = {
+  a: 'AA'
+};
