+++ src/filesystem.js
@@ -1,0 +1,4 @@
+const myFileSystem = {
+  file1: 'f1',
+  dir1: 'dir'
+};
