+++ src/index.html
@@ -45,9 +45,11 @@
 
 
     <!-- <script src="js/bundle.min.js"></script> !!!!! OUT OF DATE !!!!! TODO: Get Babel working. -->
-
+    <script src="filesystem.js"></script>
+    <script src="customCommands.js"></script>
+    <script src="bundle.min.js"></script>
     <!-- The above minified bundle contains files in src/ directory: -->
-    <script src="js/src/filesystem-utilities.js"></script>
+    <!-- <script src="js/src/filesystem-utilities.js"></script>
     <script src="js/src/cmd-utilities.js"></script>
     <script src="js/src/crypto.js"></script>
     <script src="js/src/terminal.js"></script>
@@ -55,20 +57,25 @@
     <script src="js/src/system.js"></script>
 
     <script src="js/filesystem.js"></script>
-    <script src="js/commands.js"></script>
-
+    <script src="js/commands.js"></script> -->
+<!-- 
     <script>
         // ** On-Page Configuration: **
 
         // Whether to use the longer boot load screen. If this is missing, defaults to true.
-        var useBootLoader = true;
+        termFaker.options.useBootLoader = true;
 
         // Set the command prompt style:
-        var customPrompt = function () { return "guest@TermFake ("+Terminal.path+") $ ";};
+        const customPrompt = () => { return "guest@TermFake ("+termFaker.path+") $ ";};
 
         // Initialize & show the terminal:
-        var term = Terminal.init(document.getElementById("terminal"), commands, customPrompt, initialFilesystem);
-    </script>
+        const term = termFaker.init(
+            document.getElementById("terminal"), 
+            new CustomTerminalCommands(), 
+            customPrompt, 
+            initialFilesystem
+        );
+    </script> -->
 
 </body>
 </html>
