+++ src/js/components/keysMap.js
@@ -1,0 +1,7 @@
+export const KEY_CODE_MAP = {
+  KEY_UP: 38,
+  KEY_DOWN: 40,
+  KEY_TAB: 9,
+  KEY_ENTER: 13,
+  MAX_HISTORY: 20
+};
