+++ src/js/version.js
@@ -1,0 +1,1 @@
+export const version = '1.4.5';  // Used in various commands.
