+++ src/terminal.css
